Move parseFileNameFromSymbolication to a new file called special-paths.js.
